chore: do not overwrite config by default

+++ cmd/squ/main.go
@@ -17,16 +17,17 @@ func main() {
 	config := object.DefaultConfig()
 
 	// cmd parse
-	src := flag.String("src", ".", "repo path")
-	before := flag.String("before", "HEAD~1", "before rev")
-	after := flag.String("after", "HEAD", "after rev")
-	diffOutput := flag.String("jsonOutput", "", "diff output")
-	graphOutput := flag.String("graphOutput", "", "svg output")
+	src := flag.String("src", config.SrcDir, "repo path")
+	before := flag.String("before", config.Before, "before rev")
+	after := flag.String("after", config.After, "after rev")
+	diffOutput := flag.String("jsonOutput", config.JsonOutput, "diff output")
+	graphOutput := flag.String("graphOutput", config.GraphOutput, "svg output")
 	dry := flag.Bool("dry", false, "dry")
-	indexerType := flag.String("indexer", "", "indexer type")
-	runnerType := flag.String("runner", "", "runner type")
+	indexerType := flag.String("indexer", config.IndexerType, "indexer type")
+	runnerType := flag.String("runner", config.RunnerType, "runner type")
 	sibylUrl := flag.String("sibylUrl", config.SibylUrl, "url of sibyl server")
 	debugMode := flag.Bool("debug", config.DebugMode, "debug mode switch")
+	overwriteConfig := flag.Bool("overwriteConfig", config.OverwriteConfig, "write config to file if true")
 	flag.Parse()
 
 	// load data from config file
@@ -48,11 +49,14 @@ func main() {
 	config.RunnerType = *runnerType
 	config.SibylUrl = *sibylUrl
 	config.DebugMode = *debugMode
+	config.OverwriteConfig = *overwriteConfig
 
-	bytes, err := json.MarshalIndent(config, "", "    ")
-	squ.PanicIfErr(err)
-	err = os.WriteFile(configFile, bytes, fs.ModePerm)
-	squ.PanicIfErr(err)
+	if config.OverwriteConfig {
+		bytes, err := json.MarshalIndent(config, "", "    ")
+		squ.PanicIfErr(err)
+		err = os.WriteFile(configFile, bytes, fs.ModePerm)
+		squ.PanicIfErr(err)
+	}
 
 	squ.MainFlow(config)
 }
